Move JSON parsing to respective fetcher
